cpu: rv64: refactor dwconv into unify style
