+++ src/cpu/cpu_convolution_list.cpp
@@ -76,8 +76,8 @@ using namespace dnnl::impl::cpu::aarch64;
 #elif DNNL_RV64
 #if defined(DNNL_RISCV_USE_RVV_INTRINSICS)
 #include "cpu/rv64/jit_rvv_1x1_convolution.hpp"
+#include "cpu/rv64/jit_uni_dwconv.hpp"
 #include "cpu/rv64/rvv_brgemm_conv.hpp"
-#include "cpu/rv64/rvv_dwconv.hpp"
 #include "cpu/rv64/rvv_gemm_convolution.hpp"
 #include "cpu/rv64/rvv_winograd_convolution.hpp"
 using namespace dnnl::impl::cpu::rv64;
@@ -197,7 +197,7 @@ const std::map<pk_dt_impl_key_t, std::vector<impl_list_item_t>> &impl_list_map()
             CPU_INSTANCE_AARCH64_ACL(acl_depthwise_convolution_fwd_t)
             CPU_INSTANCE_AARCH64_ACL(acl_indirect_gemm_convolution_fwd_t)
             CPU_INSTANCE_AARCH64_ACL(acl_gemm_convolution_fwd_t<f16>)
-            CPU_INSTANCE_RV64GCV(rvv_dwconv_fwd_t)
+            CPU_INSTANCE_RV64(jit_uni_dwconv_fwd_t)
             CPU_INSTANCE(ref_convolution_fwd_t)
             CPU_INSTANCE(ref_fused_convolution_fwd_t)
             nullptr,
